Clean up duplicate code and remove import cycle.
