+++ pkg/kubelet/server/stats/prometheus_resource_metrics.go
@@ -24,6 +24,77 @@ import (
 	stats "k8s.io/kubernetes/pkg/kubelet/apis/stats/v1alpha1"
 )
 
+// This file contains a series of deprecated metrics which we emit them by endpoint `/metrics/resource/v1alpha1`.
+// These metrics have been adapted to new endpoint `/metrics/resource` as well as new `Desc`s.
+// In general, we don't need to maintain these deprecated metrics any more.
+// TODO(RainbowMango): Remove this file in release 1.20.0+.
+
+var (
+	nodeCPUUsageDesc = metrics.NewDesc("node_cpu_usage_seconds_total",
+		"Cumulative cpu time consumed by the node in core-seconds",
+		nil,
+		nil,
+		metrics.ALPHA,
+		"1.18.0")
+
+	nodeMemoryUsageDesc = metrics.NewDesc("node_memory_working_set_bytes",
+		"Current working set of the node in bytes",
+		nil,
+		nil,
+		metrics.ALPHA,
+		"1.18.0")
+
+	containerCPUUsageDesc = metrics.NewDesc("container_cpu_usage_seconds_total",
+		"Cumulative cpu time consumed by the container in core-seconds",
+		[]string{"container", "pod", "namespace"},
+		nil,
+		metrics.ALPHA,
+		"1.18.0")
+
+	containerMemoryUsageDesc = metrics.NewDesc("container_memory_working_set_bytes",
+		"Current working set of the container in bytes",
+		[]string{"container", "pod", "namespace"},
+		nil,
+		metrics.ALPHA,
+		"1.18.0")
+)
+
+// getNodeCPUMetrics returns CPU utilization of a node.
+func getNodeCPUMetrics(s stats.NodeStats) (*float64, time.Time) {
+	if s.CPU == nil {
+		return nil, time.Time{}
+	}
+	v := float64(*s.CPU.UsageCoreNanoSeconds) / float64(time.Second)
+	return &v, s.CPU.Time.Time
+}
+
+// getNodeMemoryMetrics returns memory utilization of a node.
+func getNodeMemoryMetrics(s stats.NodeStats) (*float64, time.Time) {
+	if s.Memory == nil {
+		return nil, time.Time{}
+	}
+	v := float64(*s.Memory.WorkingSetBytes)
+	return &v, s.Memory.Time.Time
+}
+
+// getContainerCPUMetrics returns CPU utilization of a container.
+func getContainerCPUMetrics(s stats.ContainerStats) (*float64, time.Time) {
+	if s.CPU == nil {
+		return nil, time.Time{}
+	}
+	v := float64(*s.CPU.UsageCoreNanoSeconds) / float64(time.Second)
+	return &v, s.CPU.Time.Time
+}
+
+// getContainerMemoryMetrics returns memory utilization of a container.
+func getContainerMemoryMetrics(s stats.ContainerStats) (*float64, time.Time) {
+	if s.Memory == nil {
+		return nil, time.Time{}
+	}
+	v := float64(*s.Memory.WorkingSetBytes)
+	return &v, s.Memory.Time.Time
+}
+
 // NodeResourceMetric describes a metric for the node
 type NodeResourceMetric struct {
 	Desc    *metrics.Desc
@@ -48,6 +119,32 @@ func (n *ContainerResourceMetric) desc() *metrics.Desc {
 type ResourceMetricsConfig struct {
 	NodeMetrics      []NodeResourceMetric
 	ContainerMetrics []ContainerResourceMetric
+}
+
+// Config is the v1alpha1 resource metrics definition
+func Config() ResourceMetricsConfig {
+	return ResourceMetricsConfig{
+		NodeMetrics: []NodeResourceMetric{
+			{
+				Desc:    nodeCPUUsageDesc,
+				ValueFn: getNodeCPUMetrics,
+			},
+			{
+				Desc:    nodeMemoryUsageDesc,
+				ValueFn: getNodeMemoryMetrics,
+			},
+		},
+		ContainerMetrics: []ContainerResourceMetric{
+			{
+				Desc:    containerCPUUsageDesc,
+				ValueFn: getContainerCPUMetrics,
+			},
+			{
+				Desc:    containerMemoryUsageDesc,
+				ValueFn: getContainerMemoryMetrics,
+			},
+		},
+	}
 }
 
 // NewPrometheusResourceMetricCollector returns a metrics.StableCollector which exports resource metrics
