+++ pkg/kubelet/server/stats/prometheus_resource_metrics_test.go
@@ -25,106 +25,9 @@ import (
 	"github.com/stretchr/testify/mock"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/component-base/metrics"
 	"k8s.io/component-base/metrics/testutil"
 	statsapi "k8s.io/kubernetes/pkg/kubelet/apis/stats/v1alpha1"
-	summary "k8s.io/kubernetes/pkg/kubelet/apis/stats/v1alpha1"
 )
-
-// TODO(RainbowMango): The Desc variables and value functions should be shared with source code.
-// It can not be shared now because there is a import cycle.
-// Consider deprecate endpoint `/resource/v1alpha1` as stability framework could offer guarantee now.
-var (
-	nodeCPUUsageDesc = metrics.NewDesc("node_cpu_usage_seconds_total",
-		"Cumulative cpu time consumed by the node in core-seconds",
-		nil,
-		nil,
-		metrics.ALPHA,
-		"")
-
-	nodeMemoryUsageDesc = metrics.NewDesc("node_memory_working_set_bytes",
-		"Current working set of the node in bytes",
-		nil,
-		nil,
-		metrics.ALPHA,
-		"")
-
-	containerCPUUsageDesc = metrics.NewDesc("container_cpu_usage_seconds_total",
-		"Cumulative cpu time consumed by the container in core-seconds",
-		[]string{"container", "pod", "namespace"},
-		nil,
-		metrics.ALPHA,
-		"")
-
-	containerMemoryUsageDesc = metrics.NewDesc("container_memory_working_set_bytes",
-		"Current working set of the container in bytes",
-		[]string{"container", "pod", "namespace"},
-		nil,
-		metrics.ALPHA,
-		"")
-)
-
-// getNodeCPUMetrics returns CPU utilization of a node.
-func getNodeCPUMetrics(s summary.NodeStats) (*float64, time.Time) {
-	if s.CPU == nil {
-		return nil, time.Time{}
-	}
-	v := float64(*s.CPU.UsageCoreNanoSeconds) / float64(time.Second)
-	return &v, s.CPU.Time.Time
-}
-
-// getNodeMemoryMetrics returns memory utilization of a node.
-func getNodeMemoryMetrics(s summary.NodeStats) (*float64, time.Time) {
-	if s.Memory == nil {
-		return nil, time.Time{}
-	}
-	v := float64(*s.Memory.WorkingSetBytes)
-	return &v, s.Memory.Time.Time
-}
-
-// getContainerCPUMetrics returns CPU utilization of a container.
-func getContainerCPUMetrics(s summary.ContainerStats) (*float64, time.Time) {
-	if s.CPU == nil {
-		return nil, time.Time{}
-	}
-	v := float64(*s.CPU.UsageCoreNanoSeconds) / float64(time.Second)
-	return &v, s.CPU.Time.Time
-}
-
-// getContainerMemoryMetrics returns memory utilization of a container.
-func getContainerMemoryMetrics(s summary.ContainerStats) (*float64, time.Time) {
-	if s.Memory == nil {
-		return nil, time.Time{}
-	}
-	v := float64(*s.Memory.WorkingSetBytes)
-	return &v, s.Memory.Time.Time
-}
-
-// Config is the v1alpha1 resource metrics definition
-func Config() ResourceMetricsConfig {
-	return ResourceMetricsConfig{
-		NodeMetrics: []NodeResourceMetric{
-			{
-				Desc:    nodeCPUUsageDesc,
-				ValueFn: getNodeCPUMetrics,
-			},
-			{
-				Desc:    nodeMemoryUsageDesc,
-				ValueFn: getNodeMemoryMetrics,
-			},
-		},
-		ContainerMetrics: []ContainerResourceMetric{
-			{
-				Desc:    containerCPUUsageDesc,
-				ValueFn: getContainerCPUMetrics,
-			},
-			{
-				Desc:    containerMemoryUsageDesc,
-				ValueFn: getContainerMemoryMetrics,
-			},
-		},
-	}
-}
 
 type mockSummaryProvider struct {
 	mock.Mock
@@ -158,7 +61,7 @@ func TestCollectResourceMetrics(t *testing.T) {
 			summaryErr:           fmt.Errorf("failed to get summary"),
 			expectedMetricsNames: []string{"scrape_error"},
 			expectedMetrics: `
-				# HELP scrape_error [ALPHA] 1 if there was an error while getting container metrics, 0 otherwise
+				# HELP scrape_error [ALPHA] (Deprecated since 1.18.0) 1 if there was an error while getting container metrics, 0 otherwise
 				# TYPE scrape_error gauge
 				scrape_error 1
 			`,
@@ -185,13 +88,13 @@ func TestCollectResourceMetrics(t *testing.T) {
 				"scrape_error",
 			},
 			expectedMetrics: `
-				# HELP node_cpu_usage_seconds_total [ALPHA] Cumulative cpu time consumed by the node in core-seconds
+				# HELP node_cpu_usage_seconds_total [ALPHA] (Deprecated since 1.18.0) Cumulative cpu time consumed by the node in core-seconds
 				# TYPE node_cpu_usage_seconds_total gauge
 				node_cpu_usage_seconds_total 10 2000
-				# HELP node_memory_working_set_bytes [ALPHA] Current working set of the node in bytes
+				# HELP node_memory_working_set_bytes [ALPHA] (Deprecated since 1.18.0) Current working set of the node in bytes
 				# TYPE node_memory_working_set_bytes gauge
 				node_memory_working_set_bytes 1000 2000
-				# HELP scrape_error [ALPHA] 1 if there was an error while getting container metrics, 0 otherwise
+				# HELP scrape_error [ALPHA] (Deprecated since 1.18.0) 1 if there was an error while getting container metrics, 0 otherwise
 				# TYPE scrape_error gauge
 				scrape_error 0
 			`,
@@ -259,15 +162,15 @@ func TestCollectResourceMetrics(t *testing.T) {
 				"scrape_error",
 			},
 			expectedMetrics: `
-				# HELP scrape_error [ALPHA] 1 if there was an error while getting container metrics, 0 otherwise
+				# HELP scrape_error [ALPHA] (Deprecated since 1.18.0) 1 if there was an error while getting container metrics, 0 otherwise
 				# TYPE scrape_error gauge
 				scrape_error 0
-				# HELP container_cpu_usage_seconds_total [ALPHA] Cumulative cpu time consumed by the container in core-seconds
+				# HELP container_cpu_usage_seconds_total [ALPHA] (Deprecated since 1.18.0) Cumulative cpu time consumed by the container in core-seconds
 				# TYPE container_cpu_usage_seconds_total gauge
 				container_cpu_usage_seconds_total{container="container_a",namespace="namespace_a",pod="pod_a"} 10 2000
 				container_cpu_usage_seconds_total{container="container_a",namespace="namespace_b",pod="pod_b"} 10 2000
 				container_cpu_usage_seconds_total{container="container_b",namespace="namespace_a",pod="pod_a"} 10 2000
-				# HELP container_memory_working_set_bytes [ALPHA] Current working set of the container in bytes
+				# HELP container_memory_working_set_bytes [ALPHA] (Deprecated since 1.18.0) Current working set of the container in bytes
 				# TYPE container_memory_working_set_bytes gauge
 				container_memory_working_set_bytes{container="container_a",namespace="namespace_a",pod="pod_a"} 1000 2000
 				container_memory_working_set_bytes{container="container_a",namespace="namespace_b",pod="pod_b"} 1000 2000
@@ -283,7 +186,9 @@ func TestCollectResourceMetrics(t *testing.T) {
 			provider.On("GetCPUAndMemoryStats").Return(tc.summary, tc.summaryErr)
 			collector := NewPrometheusResourceMetricCollector(provider, tc.config)
 
-			if err := testutil.CustomCollectAndCompare(collector, strings.NewReader(tc.expectedMetrics), tc.expectedMetricsNames...); err != nil {
+			registry := testutil.NewFakeKubeRegistry("1.18.0")
+			registry.CustomMustRegister(collector)
+			if err := testutil.GatherAndCompare(registry, strings.NewReader(tc.expectedMetrics), tc.expectedMetricsNames...); err != nil {
 				t.Fatal(err)
 			}
 		})
